Codec reusability improved
