+++ src/reply.rs
@@ -4,16 +4,15 @@ use tokio::time::timeout;
 use std::{ops::Deref, pin::Pin, task::{Context, Poll}, time::Duration};
 use std::io::ErrorKind;
 use futures::{Future, SinkExt, StreamExt};
-use tokio_util::codec::{Encoder, Decoder, Framed};
+use tokio_util::codec::Framed;
 use futures::{Sink, Stream};
-use bytes::{BytesMut, Buf, BufMut};
 use pin_project::*;
 use super::options::SocketOptions;
+use super::req_reply_codec::RequestReplyCodec;
 
 const HANDSHAKE_WRITE_TIMEOUT_MS : u64 = 250;
 const REP_HANDSHAKE_PACKET : [u8; 8] = [0x00, 0x53, 0x50, 0x00, 0x00, 0x31, 0x00, 0x00];
 const REQ_HANDSHAKE_PACKET : [u8; 8] = [0x00, 0x53, 0x50, 0x00, 0x00, 0x30, 0x00, 0x00];
-
 
 pub struct NanomsgReply<S, F> {
     listener            : Option<TcpListener>,
@@ -22,8 +21,6 @@ pub struct NanomsgReply<S, F> {
     socket_options      : SocketOptions,
     _f                  : std::marker::PhantomData<F> 
 }
-
-
 
 impl <S, F>NanomsgReply<S, F>
     where S: FnMut(Vec<u8>) -> F + Clone + Send + 'static,
@@ -157,13 +154,13 @@ impl <S, F> NanomsgReplySocket <S, F>
 #[pin_project]
 struct NanomsgReplyStreamSink {
     #[pin]
-    inner           : tokio_util::codec::Framed<TcpStream, NanomsgReplyCodec>,
+    inner           : tokio_util::codec::Framed<TcpStream, RequestReplyCodec>,
 }
 
 impl NanomsgReplyStreamSink {
     fn new(tcpstream: TcpStream) -> Self {
         Self {
-            inner: Framed::new(tcpstream, NanomsgReplyCodec::new())
+            inner: Framed::new(tcpstream, RequestReplyCodec::new())
         }
     }
 }
@@ -192,7 +189,7 @@ impl <I> Sink<(u32, I)> for NanomsgReplyStreamSink
         cx: &mut Context<'_>
     ) -> Poll<Result<(), Self::Error>> {
         let this = self.project();
-        <Framed<TcpStream, NanomsgReplyCodec> as Sink<(u32, I)>>::poll_ready(this.inner, cx)
+        <Framed<TcpStream, RequestReplyCodec> as Sink<(u32, I)>>::poll_ready(this.inner, cx)
     }
 
     fn start_send(self: Pin<&mut Self>, item: (u32, I)) -> Result<(), Self::Error> {
@@ -205,7 +202,7 @@ impl <I> Sink<(u32, I)> for NanomsgReplyStreamSink
         cx: &mut Context<'_>
     ) -> Poll<Result<(), Self::Error>> {
         let this = self.project();
-        <Framed<TcpStream, NanomsgReplyCodec> as Sink<(u32, I)>>::poll_flush(this.inner, cx)
+        <Framed<TcpStream, RequestReplyCodec> as Sink<(u32, I)>>::poll_flush(this.inner, cx)
     }
 
     fn poll_close(
@@ -213,83 +210,6 @@ impl <I> Sink<(u32, I)> for NanomsgReplyStreamSink
         cx: &mut Context<'_>
     ) -> Poll<Result<(), Self::Error>> {
         let this = self.project();
-        <Framed<TcpStream, NanomsgReplyCodec> as Sink<(u32, I)>>::poll_close(this.inner, cx)
-    }
-}
-
-
-
-
-
-enum DecodingState {
-    Size,
-    RequestId(usize),
-    Payload(u32, usize)
-}
-
-struct NanomsgReplyCodec {
-    decoding_state  : DecodingState
-}
-
-impl NanomsgReplyCodec {
-    pub fn new() -> Self {
-        Self {
-            decoding_state  : DecodingState::Size,
-        }
-    }
-
-}
-
-impl <T>Encoder<(u32, T)> for NanomsgReplyCodec 
-    where T: std::ops::Deref<Target = [u8]> {
-    type Error = std::io::Error;
-
-    fn encode(&mut self, item: (u32, T), dst: &mut BytesMut) -> Result<(), Self::Error> {
-        let payload_size = item.1.len() as u64;
-        dst.put_u64(payload_size + 4);
-        dst.put_u32(item.0 | 0x80000000);
-        dst.put(item.1.as_ref());
-        Ok(())
-    }
-}
-
-
-impl Decoder for NanomsgReplyCodec {
-    type Item = (u32, Vec<u8>);
-    type Error = std::io::Error;
-
-    fn decode(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
-
-        use std::mem::replace;
-
-        loop {
-            match self.decoding_state {
-                DecodingState::Size => {
-                    if src.remaining() >= 8 {
-                        let size = src.get_u64() as usize;
-                        let _ = replace(&mut self.decoding_state, DecodingState::RequestId(size - 4));
-                    } else {
-                        return Ok(None)
-                    }
-                }
-                DecodingState::RequestId(payload_size) => {
-                    if src.remaining() >= 4 {
-                        let request_id = src.get_u32() & 0x7FFFFFFF;
-                        let _ = replace(&mut self.decoding_state, DecodingState::Payload(request_id, payload_size));
-                    } else {
-                        return Ok(None)
-                    }
-                }
-                DecodingState::Payload(request_id, payload_size) => {
-                    if src.remaining() >= payload_size {
-                        let payload = src.split_to(payload_size);
-                        let _ = replace(&mut self.decoding_state, DecodingState::Size);
-                        return Ok(Some((request_id, payload.to_vec())))
-                    } else {
-                        return Ok(None)
-                    }
-                }
-            }
-        }
+        <Framed<TcpStream, RequestReplyCodec> as Sink<(u32, I)>>::poll_close(this.inner, cx)
     }
 }
